A change has been made to the structure of the document cache and some other fixes.
Check-in By: Anshul
Bug Number: NA
Description of Checkin (fix/ enhancement): Change structure of document cache and other fixes as follow:
1. A change has been made to the structure of the document cache.
2. Changed snapticnodes.xml to application.json.
3. It is possible to load configuration from Dynamo or from a local file.
4. Dynamo DB queries implemented using repositories.
5. Both local documents and databases are based on the same model.
6. There is an addition to the message header for the originating port.

Java documentation: Y
Explanatory comments within methods: Y
Code indentation: Y
Refactoring of code: Y
Commented code removed: Y
Unnecessary print statements removed: Y

Reviewed By: Anurag Arya
Reviewer Comments: NA

+++ core/src/main/java/com/ics/id/IdCounter.java
@@ -1,7 +1,7 @@
 package com.ics.id;
 
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import com.ics.nceph.core.Configuration;
 
@@ -12,14 +12,14 @@ import com.ics.nceph.core.Configuration;
  */
 public class IdCounter
 {
-	private ConcurrentHashMap<Integer, AtomicInteger> sequenceCounters;
+	private ConcurrentHashMap<Integer, AtomicLong> sequenceCounters;
 	
 	public IdCounter() 
 	{
-		sequenceCounters = new ConcurrentHashMap<Integer, AtomicInteger>();
+		sequenceCounters = new ConcurrentHashMap<Integer, AtomicLong>();
 	}
 
-	public ConcurrentHashMap<Integer, AtomicInteger> getSequenceCounters() {
+	public ConcurrentHashMap<Integer, AtomicLong> getSequenceCounters() {
 		return sequenceCounters;
 	}
